XD-2984 Move hadoopDistro option to CommonOptions
 - Since both the `admin` and `container` use `--hadoopDistro` option
we can have this option moved to `CommonOptions`
 - Fix XD config logger to log `hadooDistro` option for all the runtime JVMs.
