+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/server/options/SingleNodeOptions.java
@@ -21,7 +21,6 @@ import javax.validation.constraints.NotNull;
 import org.kohsuke.args4j.Option;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.xd.dirt.server.options.ResourcePatternScanningOptionHandlers.HadoopDistroOptionHandler;
 import org.springframework.xd.dirt.server.options.ResourcePatternScanningOptionHandlers.SingleNodeAnalyticsOptionHandler;
 import org.springframework.xd.dirt.server.options.ResourcePatternScanningOptionHandlers.SingleNodeDataTransportOptionHandler;
 
@@ -45,10 +44,6 @@ public class SingleNodeOptions extends CommonOptions {
 
 	@Option(name = "--httpPort", usage = "HTTP port for the REST API server", metaVar = "<httpPort>")
 	private Integer httpPort;
-
-	@Option(name = "--hadoopDistro", handler = HadoopDistroOptionHandler.class,
-			usage = "The Hadoop distribution to be used for HDFS access")
-	private String distro = ContainerOptions.DEFAULT_HADOOP_DISTRO;
 
 	public Integer getPORT() {
 		return httpPort;
@@ -74,13 +69,5 @@ public class SingleNodeOptions extends CommonOptions {
 
 	public void setXD_TRANSPORT(String transport) {
 		this.transport = transport;
-	}
-
-	public void setHADOOP_DISTRO(String distro) {
-		this.distro = distro;
-	}
-
-	public String getHADOOP_DISTRO() {
-		return this.distro;
 	}
 }
